Fix cacerts file format when installing Java 11 on Debian Stretch
This is a workaround for
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=89497

A fix is applied on Debian Buster. But when installing Java 11 via
Stretch backports, the cacerts file with the PKCS12 format is not
properly generated by the ca-certificates-java package.

The workaround applied consist to migrate back to the JKS format which
is compatible with the cacerts generated by ca-certificates-java.